requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
gspread==5.12.0
oauth2client==4.1.3
pytz==2023.3
openai==1.12.0
pydub==0.25.1
python-dateutil==2.8.2
textblob==0.17.1
APScheduler==3.10.4
